(PUP-9238) Fix Puppet data file perms on Windows
Previously when storing reports on Windows it tried to set mode 640.  Due to
PUP-9106 being merged, this was causing warnings to be emitted when the owner
and group were both SYSTEM.  This is because mode 640 doesn't make sense on
Windows when both the owner and group are SYSTEM and SYSTEM should always have
Full Control (770).  This commit modifies the report storing to use mode 640
when not on Windows, and when on Windows use a mode of nil, which indicates that
the default File System permissions will be used i.e. not explicitly set.  The
permissions for this file will normally come from the Puppet Agent package which
is now set to Administrators and SYSTEM full control by default.