(PUP-9175) Use modern version of Rake gem
Previously the version of Rake was pinned to an old Ruby 1.9 compatible version
however this is beginning to cause issues with compatibility in other gems,
for example the packaging gem.  This commit changes the rake requirement in the
gemfile to use the older version for old versions of Ruby (< 2.0) and uses a
more relaxed pin, protecting on major version, for newer Ruby versions (>= 2.0).

This commit also adds a RAKE_LOCATION environment variable which can be used in
any instance to specify the the version of Rake required, much like the FACTER
and HIERA variables.

Note the rubocop disable is needed because rubocop doesn't understand the
conditional statements which ensures only one rake gem is ever actually
specified.