(maint) Reset language for every agent
We want to make sure we're resetting the language for every host. If we
have multiple servers classified as an agent, there's a potential we
reset what `language` is. So instead of `ja_JP`, we could have
`ja_JP.uft8`. This isn't so ideal on ubuntu, for example, where
`locale-gen` doesn't recognize `ja_JP.uft8` as a language, where it
would recognize `ja_JP`. Generating `ja_JP` will then fail, and the test
itself will also fail.