(maint) Avoid creating multiple log destinations
Commit dbf0ac4f3 causes puppet to only use the first log destination,
which in this test is the logfile. This broke the test as it expected
the output to be written to both the logfile and console. This commit
changes the test to just use the console log destination. The comment
about logdest overriding print debug was incorrect or at least not true
anymore.

Paired-with: Kris Bosland <kris.bosland@puppet.com>