(PUP-9252) Inject PAL Compiler also when not started via PAL
Before this if a function needed to get the compiler it would need to
know implementation details. For PAL, a feature was added to allow an
"internal function" to request injection of a "pal compiler" or a "pal
script compiler". The injection would however fail if compilation was
not initialized via PAL.

This fixes this by wrapping the internal compiler in one of the PAL
wrappers under control of the Puppet[:tasks] setting.