(PUP-9254) Add an eval() function that evaluates puppet logic in string
This adds
`function eval(String $code, Hash[String, Any] $variables) >> Any`
which will evaluate the given `$code` in a local scope where the
given variables in `$variables` are set before evaluation. The result
is returned. If any error is raised when parsing, validating or
evaluating, that will also fail the compilation.