(PUP-9132) Add configurable timeout for windows service operations
It may be useful to users to specify their own value for timing out during a
service operation. The primary use case for which is a large service (like
MySQL) that may take longer than the default timeout period to start/stop.

This commit adds a new parameter to the service type called 'timeout'.
Currently the timeout param is only available in the windows service provider