(maint) Log when non-persistent connections are closed
Previously, when creating a non-persistent http connection, we relied on
ruby to lazily start the connection and close it when the single request
was completed. Here we explicitly start and close the connection,
logging the close. Without this it appeared non-persistent connections
were never closed, though they actually were within the
`Net::HTTP.start` method[1].

[1] https://github.com/ruby/ruby/blob/v2_5_3/lib/net/http.rb#L908-L913