(PUP-5216) Add/update acceptance tests for setting Windows passwords
This commit includes the following changes:

* Add a test for setting Windows passwords independently of the attributes
property

* Add a test for setting Windows passwords while also managing the attributes
property

* Update the prevent_password_changes test to ensure that Puppet does not change
the user password if they're supposed to change their password the next time they
log-in (i.e. if the password_change_required attribute is set to true)

Note that the first test is there to enforce backwards compatibility (i.e. to
ensure that the previous behavior of setting Windows passwords is enforced when
the attributes property is not being managed). The second test is there because
implementing PUP-5216 introduced a dependency between the attributes and password
properties for Windows users; we want to ensure that this dependency results in
the correct behavior.