(PUP-9257) Make heredoc syntax check statically if possible
Before this heredoc syntax checking was only done during evaluation.
That was bad because "puppet parser validate" would then not be able to
check and report found issues.

Now, the valiator will check heredocs with syntax checker if the
resulting heredoc is a LiteralString (i.e. no interpolation). The same
check is now also made in the evaluator as there is no need to check a
literal string twice. Dynamic heredoc (with interpolation) is treated as
static if no actual interpolation was done even if marked as wanted in
the heredoc tag (i.e. `@("EOT":syntax)` but no `${}` interepolations
found.