(PUP-9270) Fix broken raised error in legacy function loader
This fixes an error in ruby_legacy_function_instantiator when it
detects that a loaded 3x function is not the expected "func_info" hash.
Instead of getting a hash, something else is returned, and the raised
error should have reported the class name of what it got instead of a
hash, but it was copy pasted from elsewhere and ended up referencing
something that was not yet instantiated.

This fixes the error so that it reports the class name of the func_info
that it got.

The problem arises when someone has mocked functions in a way that makes
them not work as functions.