(maint) Fix the AIX group provider to enforce the auth_membership param.
Prior to PUP-6542, the AIX group provider did not enforce
auth_membership, which was a bug. PUP-6542 fixed this by having the
property, _not_ the provider, enforce auth_membership. However PUP-6542
was reverted because of PUP-9265, which caused a regression.

This commit fixes the AIX group provider to enforce auth_membership. We
intend to rip this behavior out once the members property has been fixed
to enforce auth_membership at the property level instead of the provider
level in a future Puppet release (e.g. Puppet 7).