(PUP-9458) Create state machine for loading CA and CRL bundles
Implement a state machine to load CA certs and CRLs using monads. Each state
inherits from the base `State` class and implements the `next_state` method.
This allows states to be chained together. Each state performs an action, and
passes the resulting state (an SSLContext) to the next state until it either
succeeds or errors.

The machine starts in the `NeedCACerts` state and will attempt to load or
download CA certs. If successful, it will construct a new SSLContext using the
CA certs, and transition to the `NeedCRLs` state. Similarly, the `NeedCRLs`
state will attempt to load or download the CRLs, construct a new SSLContext, and
transition to the final `NeedKey` state.

The `StateMachine#ensure_ca_certificates` method returns the final SSLContext,
which can be used for subsequent connections, such as submitting a CSR.