(PUP-9268) Revert earlier merge to 5.5.x of PUP-6964
This reverts merge commit e0b7dcc5e0706534448faf0c06dc0edb42ef9b1e
where PUP-6964 was merged to 5.5.x.
We are reverting this due to backwards incompatibility loading
3x functions with illegal constructs (methods inside and outside of
the function body).

When merging this to master the revert should again be reverted.
(PUP-9137) Revert test for 4x load after 3.x load due to revert of 6964

This reverts a test that asserts that a 4.x load following a 3.x load
of a 3.x function does not again load the 3.x function. This test needs
to be reverted since PUP-6964 is reverted and it is not possible to load
3.x. functions with the 4.x loaders after that revert.

This change should not be merged to master (i.e. 6.0.z) since the
behavior of PUP-6964 and PUP-9137 should be intact there.