(PUP-9292) Change windows service query functions to take blocks
It turns out the actual query return from windows service queries was freed
before returning from teh functions, so we had a use-after-free scenario.

This commit updates the service query functions to perform the queries and
then yield a block with the query return. This way after the block completes
the return value is freed but the method would be done with that value, so no
more use-after-free