(PUP-9150) Update windows execute to pass cwd directly to Process
Since Dir.chdir is not thread safe we need to update the implementation of
Puppet::Util::Exection.execute with the cwd option. This commit updates the
implementation to pass the working directory directly to Process.create,
where Process.create will pass that directly to the API call in windows to
CreateProcess.

When we eventually move puppet to FFI calls directly to CreateProcess
(instead of using the win32-process gem) we should continue to pass the
current working dir to CreateProcess when the :cwd option is given to
Puppet::Util::Execution.execute