(maint) Add more crontab provider tests
This commit adds the following tests to the cron resource:

* A test to ensure that the crontab provider writes the crontab file of
a new user

* A test to ensure that the crontab provider fails to write the crontab
file of a nonexistent user

* A test to ensure that the crontab provider writes an originally
unauthorized user's crontab file _if_ Puppet authorizes them in the
middle of the run

These tests are part of the crontab provider's specifications and should
have been added before. They are especially useful to ensure that
PUP-9217's changes preserve backwards compatibility (especially the
first and second tests).