(PUP-9217) Return "" for some FileReadErrors in crontab filetypes
This commit modifies the crontab, suntab and aixtab filetypes to return
the empty string instead of raising a FileReadError for the following scenarios:

* Attempting to read the crontab of a nonexistent user
* Attempting to read the crontab of an unauthorized user

We do this to enforce backwards compatibility after PUP-9217's changes.
Specifically, prior to PUP-9217, the crontab provider would swallow all
FileReadErrors, choosing to treat the corresponding crontab file as
empty. This was bad in general; however, for some scenarios, treating
the crontab file as empty did make sense. These scenarios corresponded
to reading the crontab of a nonexistent user or an unauthorized user.
The former makes sense because Puppet could create the user in the
middle of the run, then proceed to create their crontab file. The latter
is an edge case and a weird scenario, but it also makes sense because
Puppet can authorize an unauthorized user to use crontab by modifying
either the cron.deny or cron.allow files in the middle of the run.

PUP-9217 no longer swallows FileReadErrors so both of these scenarios
would break. However, fixing the crontab filetypes' read methods to
return an empty string for these scenarios would preserve backwards
compatibility. It is also what should have originally been done anyways
since parsing a crontab file should be the responsibility of the
corresponding filetype instead of the provider. Thus the changes here
are considered a bug-fix even though they change the old API of the read
methods to no longer raise a FileReadError in these scenarios.