(PUP-9294) Add protection against added methods in legacy functions
Before this, methods added in a legacy function implementation file
could reside a) inside the body given to `newfunction` or b) before or
after the call to `newfunction` in the `Puppet::Parser::Functions`
module. In both a) and b) the method could be added with a `def mname`
or `def self.mname` neither defining a method where the user expects it
to be added.

This commit adds protection against these method definitions by raising
an error if found in the parsed Ruby source by using the "Ripper" to
generate a Ruby AST that is then walked.

An assertion is made that a 3x function when loaded via the
4x loaders cannot contain a method definition using either
`def name` or `def self.name` inside the function body, or outside
of it.

It will also catch other fancy meta programming trying to call
`def` or `def self` on any object.