(maint) Add example Guardfile
Automatically running the tests in a local development environment used to
be accomplished using watchr. Unfortunately, watchr hasn't been maintained
in years. This adds a Guardfile.example that can be used as a template for
a developer's own Guardfile as Guard has received updates relatively
recently as of writing this commit.