(maint) Check if user exists prior to writing their crontab
The crontab filetype passes-in the user's new crontab by creating a pipe
containing the data and passing it over to the crontab command. The
problem is that this swallows any errors associated with the crontab
command (i.e. exceptions that should be raised aren't raised). Thus,
Puppet would report that a Cron resource successfully evaluated when it,
in fact, didn't. This is bad in general, and is especially bad when the
user does not exist because in the latter case, we do want the Puppet
run to fail.

This commit adds a check for the user's existence prior to writing their
crontab. Future work should refactor the crontab filetype's write method
to at least keep it consistent with :aixtab and :suntab, or do it in a
way where the crontab command's errors aren't swallowed.