(PUP-9303) Fix problem with location of interpolation in heredoc
Before this, there were issues with the offset/length calculations for
interpolated expressions in a heredoc as the margin was not taken into
account. This obviously lead to wrong text being extracted from the
source and it caused the check for "preceding space" for `[` token
to look in the wrong position - thus possibly causing syntax errors if
that wrong position referenced whitespace.

This commit is a major change to how sublocated expressions are handled
as the earlier implementation was seriously flawed. The old
implementation used an AST element of SublocatedExpression holding
a SubLocated locator with the intention of being able to compute the
real offset/length on demand, but it managed to overwrite this
information when locating the SublocatedExpression. Thus local
offset/length was translated with the wrong locator. Further, when this
problem was addressed, the SubLocator did not perform the correct
calculation as it did correctly consider the margin. It was also
detected that even if fixed, it would not work since calculations would
need to propagate up a tree of SubLocators//real locator.

The new implementation in this commit changes this to instead mutate the
SubLocated expressions when the parser reduces this expression. Thus, a
SubLocator is only used inside of the parser and there is no longer the
need to use a SublocatedExpression.

When fixing this, a new error check is added that makes it illegal to
have a margin in a heredoc that contains text. Earlier such lines were
shifted as if they had a margin. This made it very difficult to
correctly calculate the real offset/length and thus this is now
considered illegal.

The parser_support is updated to be made aware of the intermediate state
a sublocated expression is in if there is a syntax error while parsing
it. This because the SubLocator cannot both represent local coordinates
(needed for the calculations in subspace) and global coordinates at the
same time.

Several tests are added to assert that positions are correct.