(PUP-9455) Create cert provider
Create a CertProvider with methods for loading and saving CA certs, CRLs,
client certs, private keys, and CSRs. The provider manages these in the
same way that the indirector-based implementation does:

 * All CA certs are in a single file. All CRLs are in a different file.
 * All keys, certs, requests have a name, and are loaded/saved in a
   respective directory.
 * The name is restricted to a limited set of characters, and protects
   against directory traversal.
 * The name is case-insensitive.

One important difference is that the cert provider can load keys, certs,
and requests for different hosts, such as devices, and do so without the
caller having to change global state, i.e. Puppet[:certname] = 'device'.

One change in behavior is if `:hostcert` or `:hostprivkey` are
overridden in a way that isn't consistent with `:certname`, then this
provider will not resolve the path. However, `:hostcsr` is a dead
setting so overriding it does nothing.

The provider does not yet manage file owner/group/mode permissions
during save operations, but pending tests were added.