(maint) Update Rakefile warnings task to not check generated egrammar.rb
Before this, whenever the grammar was regenerated, the warnings rake
task would find warnings in it related to indentation. Unfortunately it
is not possible to fix those without modifying racc as it produces many
lines with inconsistent indentation.

This makes the rake warnings task simply skip the generated file
puppet/pops/parser/eparser.rb