(maint) Update tests requiring a yaml end of document in output
The behavior of dropping the yaml end of document changed between
Ruby 2.4.4p296 and 2.4.5p334. The tests regarding lookup are really
not about yaml output compatibility and this commit drops
the optional end of document at the end of the output if it is present
so that the individual tests can compare against the rest of the
produced yaml output.