(PUP-9316) implement no_proxy in puppet.conf
Prior to this commit, puppet implemented http_proxy* settings in
puppet.conf or as ENV variables, if set; but only implemented the
no_proxy setting as an ENV variable.

With this commit, puppet also implements no_proxy as a setting,
and continues to implement it as an ENV variable, if set.