(PUP-9318) fix $LOAD_PATH handling in `device`
Previous to this commit, `puppet device` would use the proxy agent's
`:libdir` to load all code, since we changed the `:vardir` setting, but
not the `:libdir`. While `Puppet::Settings` has hooks to change the
`$LOAD_PATH` when `:libdir` changes, those do not trigger when the
change is effected through a interpolation, like in this case.

By setting the `:libdir` explicitely the hook is triggered, and the
`$LOAD_PATH` gets updated.

This commit also removes the `device` subcommand from the CLI's default
`$LOAD_PATH` processing.