(maint) Restructure unit tests to use built-in RSpec facilities
Rather than manually keeping track of the class that is under test, we can
tell RSpec what class we're describing, and use the built in `subject` to
get a default instance of the class, and `described_class` to refer to the
class itself.

This does not update test files that have been removed in the master
branch.