(PUP-9323) Change resolution of Deferred to be on demand
Before this, all Deferred values were resolved upfront when a catalog
was read - via a REST request to a master, or read from disk.

In PUP-9323 it is requested that the resolution of deferred values
should instead occur as late as possible to enable additional use cases;
mainly that deferred functions are guaranteed to be resolved at a time
when all of a resource's dependencies have been processed. For example,
if a resource "a" writes a file, and resource "b" depends on it, then a
deferred function call for a parameter/property in "b" will be
guaranteed to run after "a" has written the file.

This is done by modifying the Parameter and Property classes such that
the first "read" of a `value` / `should` triggers a resolution of
Deferred and Binary values (what was done upfront before this).

To support this, an instance of DeferredResolver is bound in the puppet
context by the `apply` application. This is required since resolution
takes place lazily throughout the entire apply cycle and it would be too
costly to set up a resolver per Deferred value.