(PUP-9330) Add tmpfs to SELinux filesystem whitelist
Add tmpfs as a SELinux supported filesystem type.

Currently, Puppet managed directories/files on a `tmpfs` are unable to have their SELinux attributed managed. By adding tmpfs to this array, Puppet will now be able to enforce and configure the SELinux attributes for files and directories on a tmpfs filesystem.

Signed-off-by: Jared Ledvina <jared@techsmix.net>