(PUP-5575) Fail if cannot read puppet.conf
Prior to this commit, if there were read issues with puppet.conf, either
a missing or unreadable file, we would log the issue and move on. This
isn't ideal. If puppet.conf just doesn't exist, then we should use the
defaults. However, if puppet.conf does exist but is unreadable, we
should fail. To fall back to defaults with no real warning is unexpected
and confusing behavior.