(PUP-9327) Delete CSR relative to requestdir
Previously, `puppet ssl clean` did not delete local CSRs which could lead
to the "The local CSR does not match the agent's public key" error the
next time the agent was run.

This commit modifies the `ssl` application to delete the CSR relative to
the `:requestdir` setting, instead of `:hostcsr`. The latter was last
used in the legacy SSLCertificates code circa 2011. The indirector based
implementation stores CSR in the `:requestdir` based on the `:certname`.
See commit de1986196b when the legacy implementation was removed.

This commit also completely deprecates the hostcsr to avoid future
confusion.