(maint) Set server in main section of conf
The puppetserver ca cli tool needs the value of the server set in "main"
or "master" sections of the puppet.conf to function correctly.

Previously we updated the main portions of the autosign and certificate
extension tests to set server in the master section and removed our
passing of --server to the agent.

The agent however will not see the server setting in the master
section. This moves the server setting to the main section. It
additionally wraps the teardown calls to puppetserver ca in a
with_puppet_running to with server set.