(maint) Make the log expectation in logging_spec.rb more flexible
While performing a 5.5.x => 6.0.x merge-up, it was found that the
Appveyor image would report a nil value for the domain fact. Facter
would log this as an additional debug message so that the log size
assertion would fail on Appveyor.

This commit replaces the log expectations with a single expectation
that checks if the test-generated log message is included in the logs.
This way, the test still tests the right thing while allowing for some
flexibility in the generated log messages in case e.g. Facter generates
more debug messages or we have additional log messages generated from
other sources.