(PUP-9247) Check process name when clearing pidlock
Under rare but possible conditions, a stale Puppet pidlock file may
exist which contains a pid that refers to another (non-puppet) process.
This adds a check of the process name to avoid locking issues under these
circumstances.

Due to the complexities of cross-platform process management, this
additional check is currently limited to POSIX platforms only.

The unit tests are disabled on JRuby due to the use of
Puppet::Util::Execution, which does a fork and exec on posix systems, but
fork is not supported in JRuby