(PUP-9329) Fix short form format rule for Array/Hash in string converter
Before this, if specifying a short format (e.g. "%#a") for an
Array or similar for a Hash, the rule would not be applied as
a collection format.

This commit adds the special processing to take the simplified form
when given for a top level Array or Hash to be applied as a format
specification of `{ BASE_TYPE => { 'format' => GIVEN }`.