(PUP-9319) Use remote environment for `puppet device`
This commit uses a "remote" environment, instead of a full one. This
allows the override to work, even if we do not have that environment's
codedir available locally.

In the case of a full puppet run, the `Configurer` does additional work
to update and sync the plugindir vs what the ENC specifies.

# Conflicts:
#	spec/unit/application/device_spec.rb