(PUP-8687) Add support for logging with JSON lines output format
Puppet had logging to JSON for some time, but the generated log file is not a
valid JSON document because of the way a log file is written to.  [JSON lines]
(aka newline-delimited JSON) format is a text file format that use newlines to
separate JSON documents, making it easy to process entries one at a time, and
therefore not suffering from this deficiency.

Convention suggests that [JSON lines] files may be saved with the file
extension .jsonl.

[JSON lines]: http://jsonlines.org/