(PUP-9356) Allow loading local content for local `puppet device` modes
The changes in PUP-8766/#7242 removed plugin loading from a local
environment for `puppet device`, even in the cases where local content
is supposed to be used.

This change reverts the part of #7242 that was causing this. By allowing
puppet to load local content for '--apply', '--facts' and '--resource',
those run modes can now work stand alone again.