(PUP-9359) Fix error when 3x function contains calls not based on id
Before this, when the ruby_legacy_function_loader loaded a 3x function
it scanned the ruby parse tree for calls for the purpose of finding
a call to "newfunction". When doing so, the code did not anticipate
that there are other forms of calls where there is no method name e.g.
Float("3.14"). This resulted in return of `nil` from find_identity.
A subsequent lookup of [1] in that array failed.

This is now fixed by checking if there is an identity (i.e. method
name). And if not, it cannot possibly be the expected "newfunction"
call.

A unit test is added that a Float("3.14") works inside a 3x function.