(PUP-9303) Fix problem with not being bug compatible for outside margin
Before this, the fix of the heredoc locator issues in PUP-9303 made
the new implementation incompatible with the earlier behavior where
a heredoc with margin specification and lines with text in the margin
was simply kept verbatim (no margin adjustment).

That behavior is in fact also a bug since the specification states that
whitespace up to the margin should be removed.

This makes the fix for PUP-9303 bug compatible with the previous version
while also fixing the problem described in PUP-9303.