(maint) Bump minimum Facter gem version & remove obsolete tests
Facter 2.4.0 has been out since Jan 22, 2015, and is the minimum version
to contain all of the Facter.debuggin, Facter.trace, and Facter.on_message
methods. By requiring at least version 2.4.0 of the Ruby Facter, we no
longer have to carry along tests with fragile setup to test for Puppet
working if the version of Facter is too old to have any of these methods.