(PUP-9336) Allow Boolean confines to accept a lambda as a value
Currently, any value passed to the Boolean true/false confines are
evaluated when the provider's loaded. This is OK for the most part;
however, there are scenarios where lazy evaluation may be useful. By
"lazy evaluation", we mean evaluating the confine's value when checking
for provider suitability since that is where the value is needed.

This commit extends the Boolean confines to support lazy evaluation by
allowing one to pass-in a lambda for the confine's value. For example,
the changes in this commit would support something like:

confine :true => lambda do
  true
end

which is semantically equivalent to:

confine :true => true