(PUP-9336) Lazily evaluate the upstart daemon check confine
Previously, the upstart daemon check confine was evaluated when the
upstart provider was loaded. This caused issues in running the unit
tests on JRuby for Puppet 6, and it also meant that the upstart provider
could not become suitable during the course of a transaction. The latter
scenario is possible if one installs the upstart package and then uses the
upstart provider in the same Puppet run.

This commit modifies the upstart daemon check confine to be lazily
evaluated, which should fix the issue.