(PUP-9364) Exclude unnecessary module release metadata
The `puppet module install` command will download release metadata[1]
for each release of the module and its dependencies so that the PMT can
resolve dependencies locally. The v3 'releases' forge API response is
verbose as the release metadata includes the module's readme, changelog
and license, which are not needed to resolve dependencies.

This commit excludes all currently known fields except version,
metadata, and file_md5 to reduce the amount of transferred data. Ideally
we could specify only the fields we want to be included, but the
`include_field` parameter doesn't work that way.

For stdlib specifically, this PR drops the amount of data transferred
from 2.5MB to 77KB. This will also benefit modules that depend on
stdlib.

[1] https://forgeapi.puppet.com/#!/release/getReleases