(PUP-6488) Allow alternate package commands, do not depend upon PATH.
With this commit ...

The package type and provider accept a 'target' parameter / feature,
allowing for management of packages when multiple installations of
a package management command are present on a system, but not in the PATH.

The gem, pip, and pip3 providers implement the target parameter / feature.

Changes to the gem also addresses PUP-6134.