(PUP-9294) Add protection against possible future changes in Ripper
Before this, the extraction of line information from the "rast" produced
by Ripper did not check if the expected information was in an Array or
not, and it simply extracted a position.

This commit adds protection against this.
Note that, if Ripper changes, tests will fail because information will
be missing. At runtime, the expected errors will be raised instead of a
cryptic failure to obtain the line information.