(PUP-9389) Add staging_location parameter to File
When Puppet creates a new file it creates it in a temporary location.
This location is currently the location the file will exist in plus a
unique name, eg /etc/foo123fsd for /etc/foo. In cases where a user
wants to first render a file to a temporary staging location, perform
some validation commands and then move the file to the final location.

This avoids issues where, for example, an application reads all the
files in a directory when performing a task (sudo) or an
application reloads configs automatically from a directory.