(PUP-9399) Pass srv domain to resolver for CA routes
SRV resolution was broken for CA related routes because we were passing
the `ca` SRV service name as the SRV domain, and ignoring the
`srv_domain` setting. This wasn't noticed because the service name will
default to `:puppet`.

This commit passes the SRV domain name as the first argument and service
name as the second argument.