(PUP-4543) Update acceptance readme details
It's been a few years since we've overhauled details on how to run
puppet acceptance. This commit completely rewrites this file to update
the instructions. A lot of the methods we use have changed or updated
since this file was written.

This update includes some additional new details, like running tests
ith docker and from a git checkout.