(PUP-9336) Return false on jRuby in upstart daemon check confine
Puppet::Util::Execution.execute does not currently work on jRuby.
Unfortunately, since this confine is invoked whenever we check for
provider suitability and since provider suitability is still checked
on the master, this confine will still be invoked on the master. Thus
to avoid raising an exception, we do an early return if we're running
on jRuby.