(PUP-6885) Remove server and server_list conflict warning
From the ticket description, when we added `server_list`, it was with
the idea that it would be used in place of `server` for all things. We
never really made strides with that, and now have a mixture of code that
uses `server` or `server_list` or just the first entry from
`server_list`.

Generally, `agent` is the only puppet application that knows about both
`server` and `server_list`.