(PUP-9303) Fix problem with SubLocatedExpression no longer in use
With PUP-9303 the handling of heredoc is changed so that a
SubLocatedExpression is no longer needed as a wrapper around the
contents of a Heredoc. PUP-9303 will instead recalculate all offsets
in expressions inside of a heredoc during parsing.

When the revert of reverting PUP-9303 was done, a new feature that
validates heredoc syntax was introduced. A static heredoc (no
interpolation) is validated statically, and heredoc with interpolation
is validated by the evaluator.

Both those locations assumed that a Heredoc always contains a
SubLocatedExpression - and this is no longer true when PUP-9303 logic
was again applied.

This commit fixes both locations by simply using the Heredoc's
`text_expr` without getting the `expr` from `SubLocatedExpression`.