(PUP-9357) Give debug message with command when exec fails check
When we abort an exec because of a failed check,
log a debug level message with the command since that information is not easily available.
This change also enables :unless and :onlyif to be marked sensitive, since it seems
like they would have similar reasons as the main command to be redacted.