(PUP-9477) Make call() load any function
Before this, the call() function looked up its own loader and used that
to search for the referenced function to call. That only worked when
calling functions in puppet core since the system loader only searches
itself and the parent static loader.

The change is to treat the call as an external call thus searching all
of the loaders from the perpective of the calling scope.