(PUP-9025) Remove filebucket type server/port defaults
The fileucket type should not have a default set for server or port. We
have logic built in to find both of those pieces of information if they
aren't explicitely set in the resource manifest. We should be using that
default logic rather than arbitrarily setting these defaults for the user.

This wouldn't be in such a compicated state currently if it weren't for
the confusion between `server` and `server_list`. Right now, `puppet agent`
and friends prefer `server_list` over server. The filebucket application
also prefers `server_list` over `server`. If `server_list` is not set,
we should use `server`.

Removing these defaults, we get closer parity between how the filebucket
application and the filebucket resource behave when picking what server
to hit.