(PUP-9455) Make file unwritable on Windows
If a directory is non-traversable, then Windows will not prevent its
children from being written/deleted due to the "Bypass traverse
checking" user privilege, which is enabled by default. To make a file
unwritable, we can open a handle to the file, which will prevent other
processes, including ourselves from overwriting it.