(PUP-9496) Allow MSI packages to install without setting CWD
After PUP-6920, the Windows package provider requires the "directory"
part of the source path to be something which can be set to the
working directory. Unfortunately, for http sources of MSI files, chdir
doesn't work. This has broken installing those packages.

This change allows MSIs to be installed without setting the working
directory if it does not exist. It still prefers to set the directory
when possible, just in case any MSIs out there are as badly-behaved as
legacy exe installers are.