(PUP-9337) Update windows file provider to treat SYSTEM group/user as in sync
In situations on windows where the user is not managing the group/user and one
or both is set to the SYSTEM user: Puppet should treat the resource as
"insync" if the system permissions are FullControl, and out of sync otherwise
regardless of what the requested permissions are.

When not explicitly managing the group/user SYSTEM should always be
FullControl.