(PUP-9337) Update puppet to warn when managing system and requesting low perms
If user's are managing the SYSTEM user in file permissions, they should not be
setting permissions for SYSTEM to less than Full Control. Anything less than
that may have unintended consequences for files.

This commit adds warnings for the specific scenario where a user is explicitly
managing the SYSTEM user in the user or group parameters and setting
permissions to less than full control.