(PUP-9456) Create SSL Provider
Creates an SSLProvider with methods for creating SSLContext objects. In the
future, it will be possible to create TLS connections from an SSLContext.

Connections created using `create_insecure_context` will use `VERIFY_NONE` and
will not be authenticated. Do not call this method.

Connections created using `create_root_context` will verify the server (both
verifying the certificate chain and verifying the server's certificate matches
the host it tried to connect to), but the server will not verify the client.
This is useful in situations where the client doesn't yet have a client cert.

Connections created using `create_context` will be mutually verified, provided
the client requests or requires a client cert. If not, then the connection will
behave the same as `create_root_context`.

The `create_context` method ensures it can construct a valid cert chain from the
client cert to the root. If the client cert is expired, invalid, revoked, an
intermediate or root CA is missing, doesn't have the proper extensions, etc,
then an exception will be raised. If the revocation mode is `:chain`, then there
must be a CRL for each CA cert. If the mode is `:leaf`, then there must be a CRL
for the intermediate CA that issued the peer's certificate. If `false`, then no
CRL validation will be performed.